# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Emerj Pro API",
        "description": "API Documentation of Emerj Pro System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/paymentMethods": {
            "get": {
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Get a list of payment methods",
                "operationId": "155b93c619012d0259cd07cecd01eff6",
                "responses": {
                    "200": {
                        "description": "List of payment methods",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "payment_method": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Register a new payment method",
                "operationId": "10d63b68f35f59816f607924e39f6284",
                "requestBody": {
                    "description": "Payment method data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "payment_method_type_id",
                                    "name",
                                    "last_digits",
                                    "token"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer"
                                    },
                                    "payment_method_type_id": {
                                        "description": "Payment method type ID",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Name of the payment method",
                                        "type": "string"
                                    },
                                    "last_digits": {
                                        "description": "Last digits of the payment method",
                                        "type": "string"
                                    },
                                    "token": {
                                        "description": "Payment token",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment method added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment method added successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "payment_method": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt_auth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}